+++ .storybook/main.template.js
@@ -1,7 +1,7 @@
 module.exports = {
   "stories": [
-    "../playground/src/**/*.stories.mdx",
-    "../playground/src/**/*.stories.@(js|jsx|ts|tsx)",
+    "../src/**/*.stories.mdx",
+    "../src/**/*.stories.@(js|jsx|ts|tsx)",
   ],
   "addons": [
     "@storybook/addon-docs",